Program Magang
This project is made to monitoring the Humidity and Temperature of server rooms in PT Bumi Menara Internusa.
There are 3 rooms of server in this company so, it must be hard to monitor the temperature of server rooms by EDP, Electronic Data Processing, members because the places of server room separate by apart.
In this project we made the EDP Members easily to monitor the temperature using the Web Site. Data Serial, RS485 Modul, used to communicate between embedded device in room 1 and room 2. Ethernet Shield used to communicate between Embedded device in room 1 and Central of Server.